Refine(autofill): Modernize AutoFill interface
- Based on collection view with a diffable data source
- Entry list can be navigated with keyboard and Cmd+F
- In field selection mode, expands the entry cell instead of showing popup menu
- Adds "Copy %field% to clipboard" context menu for entries
- AutoFill Coordinator and Entry Finder are refactored into multiple files
- Removes "Search" and "Lock DB" buttons as useless

Known issue: on macOS 15.6, in AutoFill only, pressing Down arrow from the search field does not transfer focus to the list. Seems to be a system bug in forwarding events without an underlying window. (If used in the main app, the same code works just fine.) Workaround: press Enter in search field to start selecting items.